Factor the general method out
